major update for directory solution
